move initial raycast impl test into wand listener
