+++ src/main/java/dev/tizu/headmate/ThisPlugin.java
@@ -6,6 +6,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 import dev.tizu.headmate.editor.EditorListener;
 import dev.tizu.headmate.headmate.HeadmateListener;
 import dev.tizu.headmate.menu.Menu;
+import dev.tizu.headmate.wand.WandListener;
 
 public class ThisPlugin extends JavaPlugin {
     public static ThisPlugin instance;
@@ -16,5 +17,6 @@ public class ThisPlugin extends JavaPlugin {
         Bukkit.getPluginManager().registerEvents(new EditorListener(), this);
         Bukkit.getPluginManager().registerEvents(new HeadmateListener(), this);
         Bukkit.getPluginManager().registerEvents(new Menu.MenuListener(), this);
+        Bukkit.getPluginManager().registerEvents(new WandListener(), this);
     }
 }
